spline #954 TxBuilder refactoring
