+++ persistence/src/main/scala/za/co/absa/spline/persistence/migration/Migrator.scala
@@ -84,7 +84,7 @@ class Migrator(
       )
       _ <- db.collection(DBVersion.name).insertDocument(model.DBVersion(version.asString, Status.Preparing)).toScala
       _ <- db.adminExecute(script)
-      _ <- new TxBuilder()
+      _ <- new JSTxBuilder()
         .addQuery(UpdateQuery(DBVersion,
           s"${UpdateQuery.DocWildcard}.status == '${Status.Current}'", Map("status" -> Status.Upgraded.toString)))
         .addQuery(UpdateQuery(DBVersion,
